Added quotes
ERROR! Syntax Error while loading YAML.
  found unacceptable key (unhashable type: 'AnsibleMapping')

The error appears to be in '/Users/psuedofinnish/.ansible/roles/ansible-keycloak/tasks/setup_lets_encrypt.yml': line 9, column 27, but may
be elsewhere in the file depending on the exact syntax problem.

The offending line appears to be:

  vars:
    certbot_admin_email: {{ letsencrypt_certbot_admin_email }}
                          ^ here
We could be wrong, but this one looks like it might be an issue with
missing quotes. Always quote template expression brackets when they
start a value. For instance:

    with_items:
      - {{ foo }}

Should be written as:

    with_items:
      - "{{ foo }}"